+++ src/main/java/app/elements/JoinHandler.java
@@ -1,0 +1,5 @@
+package app.elements;
+
+public interface JoinHandler {
+  void clickedJoin(String lobbyId);
+}