+++ src/main/java/protocol/LobbyModel.java
@@ -9,6 +9,8 @@ public class LobbyModel {
 
   public String id;
 
+  public String name;
+
   public List<String> players;
 
   public boolean isRunning;
@@ -18,11 +20,13 @@ public class LobbyModel {
    *
    * @param id lobby id
    * @param players connected players
+   * @param name lobby name
    * @param isRunning whether or not a race is currently running
    */
-  public LobbyModel(String id, List<String> players, boolean isRunning) {
+  public LobbyModel(String id, List<String> players, String name, boolean isRunning) {
     this.id = id;
     this.players = players;
+    this.name = name;
     this.isRunning = isRunning;
   }
 }